wait_time_display/requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
